moved content to utils file
